\name{run4.show.prediction}
\Rdversion{1.1}
\alias{run4.show.prediction}
\title{
Plot predicted values and ROC curve
}
\description{
Creates a plot of the predicted values and the corresponding ROC curve for the resulting prediction file generated by either \code{\link{run3.prediction.train.test}} or \code{\link{run2.prediction.cvv}} functions.
}
\usage{
run4.show.prediction(filename)
}
\arguments{
  \item{filename}{
	The file produced either by \code{\link{run2.prediction.cvv}} or \code{\link{run3.prediction.train.test}} functions. 
}
}
\value{
\item{out.tpr}{The true positive rate.}
\item{out.fpr}{The false positive rate.}
\item{out.rocarea}{The ROC curve area.}
}
\author{
Laurent Briollais, Adrian Dobra, Olga Vesselova
}
\note{
This function takes a while to run. 
}

\seealso{
\code{\link{run4.save.prediction}}, \code{\link{run1.moss.regression}}, 
\code{\link{run2.prediction.cvv}}, \code{\link{run3.prediction.train.test}}

}
\examples{

fname <- "randbinary.txt"
write(rbinom(200,1,0.5), file=fname, append=FALSE, sep=" ", ncolumns=50)
name.reg <- run1.moss.regression(fname)
name.cvv <- run2.prediction.cvv(fname, name.reg)
run4.show.prediction(name.cvv)

try(system(paste("rm ", fname, "*", sep="")))

}
\keyword{hplot}
\keyword{misc}
