\name{simuCC}
\Rdversion{1.1}
\alias{simuCC}
\docType{data}
\title{A simulated case-control sample.}
\description{The data was simulated using code from the simuPOP cookbook on simupop.sourceforge.net. It is a sample of 1000 cases and 1000 controls from a simulated but realistic population. It contains the genotype information at 6000 diallelic SNPs (i.e. SNPs with three categories) and the disease status for each individual (which is the last column). The markers 'rs4491689' and 'rs6869003' are associated with the disease.}
\usage{data(simuCC)}
\source{The python code for simulating the data is example2.py on http://simupop.sourceforge.net/cookbook/pmwiki.php/Cookbook/SimuGWAS. See the references for more information on how the data were generated.} 
\references{[1] Peng, B. and Amos, C. (2010). Forward-time simulation of realistic samples for genome-wide association studies. BMC Bioinformatics, 11, 442-453.  

[2] Peng, B. and Kimmel, M. (2005). simuPOP: a forward-time population genetics simulation environment. Bioinformatics, 21, 3686-3687.}
\keyword{datasets}
